moves protocol related classes to different project
