--- ingest/src/main/scala/milo/device/Measurement.scala
@@ -1,3 +1,0 @@
-package milo.device
-
-trait Measurement
