+++ ingest/src/test/scala/milo/server/DeviceEmulator.scala
@@ -10,12 +10,16 @@ import org.scalacheck._
  * Sends generated data over tcp connection to the data processor on the
  * other end. Should suffice the following requirements:
  *   - Be configurable
- *   - Be controlable (send one message, multiple generated messages)
+ *   - Be controllable (send one message, multiple generated messages)
  */
 final class DeviceEmulator(socket: InetSocketAddress) extends Actor {
   import context.system
 
   IO(Tcp) ! Tcp.Connect(socket)
+
+  def receive: Receive = {
+    case _ =>
+  }
 
 }
