+++ protocols/build.sbt
@@ -1,1 +1,5 @@
+libraryDependencies ++= Seq(
+  "com.typesafe.akka" %% "akka-actor"   % "2.3.12"
+)
+
 com.trueaccord.scalapb.ScalaPbPlugin.protobufSettings
