+++ protocols/src/main/scala/platform/device/DeviceConfiguration.scala
@@ -1,4 +1,4 @@
-package milo.device
+package platform.device
 
 trait DeviceConfiguration {
