+++ protocols/src/main/scala/platform/device/DeviceData.scala
@@ -1,4 +1,4 @@
-package milo.device
+package platform.device
 
 trait DeviceData{
   def time: Long
