+++ protocols/src/main/scala/platform/device/Measurement.scala
@@ -1,0 +1,3 @@
+package platform.device
+
+trait Measurement
